+++ ArtInk/ArtInk.Infraestructure/Repository/Interfaces/IRepositoryFactura.cs
@@ -10,7 +10,7 @@ namespace ArtInk.Infraestructure.Repository.Interfaces
     public interface IRepositoryFactura
     {
         Task<ICollection<Factura>> ListAsync();
-        Task<Factura?> FindByIdAsync(int id);
+        Task<Factura?> FindByIdAsync(long id);
 
     }
 }
